requirements.txt
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
discord-protos==0.0.2
discord.py-self @ git+https://github.com/dolfies/discord.py-self
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
protobuf==6.33.4
Pygments==2.19.2
requests==2.32.5
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
